refactor: extract bash access classification
