+++ crates/tools/src/lib.rs
@@ -39,36 +39,6 @@ impl BashTool {
     }
 }
 
-fn classify_bash_fragment(fragment: &str) -> alan_agent_protocol::ToolCapability {
-    if fragment.is_empty() {
-        return alan_agent_protocol::ToolCapability::Read;
-    }
-
-    let tokens: Vec<&str> = fragment.split_whitespace().collect();
-    if tokens.is_empty() {
-        return alan_agent_protocol::ToolCapability::Read;
-    }
-    let effective_tokens = effective_command_tokens(&tokens);
-    let effective_tokens = effective_tokens.as_slice();
-
-    if contains_unsupported_shell_form(&tokens) {
-        return alan_agent_protocol::ToolCapability::Unknown;
-    }
-    if is_network_command(fragment, effective_tokens) {
-        return alan_agent_protocol::ToolCapability::Network;
-    }
-    if contains_nested_eval_wrapper(&tokens) {
-        return alan_agent_protocol::ToolCapability::Unknown;
-    }
-    if is_write_command(fragment, effective_tokens) {
-        return alan_agent_protocol::ToolCapability::Write;
-    }
-    if is_safe_read_command(effective_tokens) || is_wrapper_query_command(&tokens) {
-        return alan_agent_protocol::ToolCapability::Read;
-    }
-    alan_agent_protocol::ToolCapability::Unknown
-}
-
 fn contains_unsupported_shell_form(tokens: &[&str]) -> bool {
     let Some(command_index) = tokens.iter().position(|word| !is_env_assignment(word)) else {
         return false;
@@ -82,310 +52,11 @@ fn contains_unsupported_shell_form(tokens: &[&str]) -> bool {
     is_unsupported_shell_wrapper(command_basename(command_word))
 }
 
-fn is_network_command(fragment: &str, tokens: &[&str]) -> bool {
-    // Match on the basename so path-qualified forms (`/usr/bin/curl`) classify
-    // like the bare head; otherwise an approved network call would run with the
-    // sandbox network deny still in force and fail.
-    let head = command_basename(tokens[0]);
-    if matches!(
-        head,
-        "curl" | "wget" | "ssh" | "scp" | "sftp" | "nc" | "netcat" | "socat" | "telnet" | "ftp"
-    ) {
-        return true;
-    }
-
-    let pair = tokens.get(1).copied().unwrap_or_default();
-    if (head == "git" && is_git_network_command(tokens))
-        || (head == "docker" && pair == "pull")
-        || (head == "npm" && pair == "install")
-        || (head == "pnpm" && pair == "add")
-        || (head == "yarn" && pair == "add")
-        || ((head == "pip" || head == "pip3") && pair == "install")
-        || (head == "cargo" && pair == "install")
-        || (head == "brew" && pair == "install")
-        || ((head == "apt" || head == "apt-get" || head == "yum" || head == "dnf")
-            && pair == "install")
-    {
-        return true;
-    }
-
-    // Catch explicit http(s) fetch commands wrapped in generic shells.
-    fragment.contains("http://") || fragment.contains("https://")
-}
-
-fn is_write_command(fragment: &str, tokens: &[&str]) -> bool {
-    // Match on the basename so path-qualified forms (`/bin/rm`) classify like the
-    // bare head instead of falling through to Unknown.
-    let head = command_basename(tokens[0]);
-    if matches!(
-        head,
-        "rm" | "rmdir" | "mv" | "cp" | "chmod" | "chown" | "mkdir" | "touch" | "truncate"
-    ) {
-        return true;
-    }
-
-    if head == "sed" && sed_in_place_flag(tokens) {
-        return true;
-    }
-
-    if head == "find" && find_has_write_action(tokens) {
-        return true;
-    }
-
-    if is_local_verification_command(tokens) {
-        return true;
-    }
-
-    if head == "git" {
-        if is_git_network_command(tokens) {
-            return false;
-        }
-        if !is_git_read_command(tokens) {
-            return true;
-        }
-    }
-
-    contains_output_redirection(fragment)
-}
-
-fn sed_in_place_flag(tokens: &[&str]) -> bool {
-    tokens.iter().skip(1).copied().any(|token| {
-        token == "-i"
-            || token == "--in-place"
-            || short_option_cluster_contains_flag(token, 'i')
-            || token
-                .strip_prefix("-i")
-                .is_some_and(|suffix| !suffix.is_empty())
-            || token.starts_with("--in-place=")
-    })
-}
-
-fn short_option_cluster_contains_flag(token: &str, flag: char) -> bool {
-    token.starts_with('-') && !token.starts_with("--") && token.chars().skip(1).any(|ch| ch == flag)
-}
-
-fn find_has_write_action(tokens: &[&str]) -> bool {
-    tokens.iter().skip(1).copied().any(|token| {
-        matches!(
-            token,
-            "-exec"
-                | "-execdir"
-                | "-delete"
-                | "-ok"
-                | "-okdir"
-                | "-fprint"
-                | "-fprint0"
-                | "-fprintf"
-                | "-fls"
-        )
-    })
-}
-
-fn is_local_verification_command(tokens: &[&str]) -> bool {
-    let head = command_basename(tokens[0]);
-    let pair = tokens.get(1).copied().unwrap_or_default();
-
-    if matches!(head, "tox" | "nox") {
-        return !is_tool_query_command(tokens);
-    }
-
-    if matches!(head, "pytest" | "py.test" | "nosetests" | "nosetests3") {
-        return true;
-    }
-
-    if (head == "cargo" && matches!(pair, "test" | "check" | "clippy"))
-        || (head == "go" && pair == "test")
-        || (matches!(head, "npm" | "pnpm" | "yarn" | "bun") && pair == "test")
-        || (matches!(head, "make" | "just") && matches!(pair, "test" | "check"))
-    {
-        return true;
-    }
-
-    if python_module_command(tokens).is_some_and(|module| matches!(module, "pytest" | "unittest")) {
-        return true;
-    }
-
-    local_verification_subject(tokens)
-        .is_some_and(|(command, args)| is_local_verification_entrypoint(command, args))
-}
-
-fn is_python_query_command(tokens: &[&str]) -> bool {
-    let head = command_basename(tokens.first().copied().unwrap_or_default());
-    if !matches!(head, "python" | "python3") {
-        return false;
-    }
-
-    is_tool_query_command(tokens)
-}
-
-fn is_tool_query_command(tokens: &[&str]) -> bool {
-    tokens
-        .iter()
-        .skip(1)
-        .copied()
-        .find(|token| !token.is_empty())
-        .is_some_and(|arg| matches!(arg, "-h" | "--help" | "--version") || arg.starts_with("-V"))
-}
-
-fn python_module_command<'a>(tokens: &'a [&'a str]) -> Option<&'a str> {
-    let head = command_basename(tokens.first().copied()?);
-    if !matches!(head, "python" | "python3") {
-        return None;
-    }
-
-    let mut index = 1;
-    while let Some(token) = tokens.get(index).copied() {
-        if token == "-m" {
-            return tokens.get(index + 1).copied();
-        }
-        if !token.starts_with('-') {
-            return None;
-        }
-        index += 1;
-    }
-
-    None
-}
-
-fn local_verification_subject<'a>(tokens: &'a [&'a str]) -> Option<(&'a str, &'a [&'a str])> {
-    let command = tokens.first().copied()?;
-    if is_local_command_path(command) {
-        return Some((command, &tokens[1..]));
-    }
-
-    python_script_command(tokens)
-}
-
-fn python_script_command<'a>(tokens: &'a [&'a str]) -> Option<(&'a str, &'a [&'a str])> {
-    let head = command_basename(tokens.first().copied()?);
-    if !matches!(head, "python" | "python3") {
-        return None;
-    }
-
-    let mut index = 1;
-    while let Some(token) = tokens.get(index).copied() {
-        if token == "-m" || token == "-c" || token == "-" {
-            return None;
-        }
-        if !token.starts_with('-') {
-            return Some((token, &tokens[index + 1..]));
-        }
-        index += 1;
-    }
-
-    None
-}
-
-fn is_local_verification_entrypoint(command: &str, args: &[&str]) -> bool {
-    if !is_local_command_path(command) {
-        return false;
-    }
-
-    if is_verification_entrypoint_name(command_basename(command)) {
-        return true;
-    }
-
-    first_non_option_arg(args).is_some_and(is_verification_subcommand)
-}
-
-fn is_local_command_path(command: &str) -> bool {
-    if command.is_empty() {
-        return false;
-    }
-
-    if command.starts_with("./") || command.starts_with("../") || command.contains('/') {
-        return true;
-    }
-
-    matches!(
-        Path::new(command).extension().and_then(|ext| ext.to_str()),
-        Some("py" | "sh" | "rb" | "pl" | "php" | "js")
-    ) || matches!(command_basename(command), "gradlew" | "mvnw")
-}
-
-fn is_verification_entrypoint_name(command: &str) -> bool {
-    let stem = Path::new(command)
-        .file_stem()
-        .and_then(|name| name.to_str())
-        .unwrap_or(command)
-        .to_ascii_lowercase();
-
-    matches!(
-        stem.as_str(),
-        "runtests" | "run-tests" | "run_tests" | "pytest" | "nosetests" | "nosetests3"
-    ) || stem == "test"
-        || stem == "tests"
-}
-
-fn first_non_option_arg<'a>(args: &'a [&'a str]) -> Option<&'a str> {
-    let mut saw_double_dash = false;
-    for arg in args {
-        if arg.is_empty() {
-            continue;
-        }
-        if !saw_double_dash {
-            if *arg == "--" {
-                saw_double_dash = true;
-                continue;
-            }
-            if arg.starts_with('-') {
-                continue;
-            }
-        }
-        return Some(*arg);
-    }
-
-    None
-}
-
-fn is_verification_subcommand(arg: &str) -> bool {
-    matches!(arg, "test" | "tests" | "check" | "clippy" | "verify")
-}
-
 fn contains_nested_eval_wrapper(tokens: &[&str]) -> bool {
     let Some(view) = nested_eval_command_view(tokens) else {
         return false;
     };
     view.opaque_wrapper || leading_eval_flag(view.command, view.args).is_some()
-}
-
-fn contains_output_redirection(fragment: &str) -> bool {
-    let mut in_single = false;
-    let mut in_double = false;
-    let mut escaped = false;
-
-    for ch in fragment.chars() {
-        if escaped {
-            escaped = false;
-            continue;
-        }
-
-        if in_single {
-            if ch == '\'' {
-                in_single = false;
-            }
-            continue;
-        }
-
-        if in_double {
-            match ch {
-                '\\' => escaped = true,
-                '"' => in_double = false,
-                _ => {}
-            }
-            continue;
-        }
-
-        match ch {
-            '\\' => escaped = true,
-            '\'' => in_single = true,
-            '"' => in_double = true,
-            '>' => return true,
-            _ => {}
-        }
-    }
-
-    false
 }
 
 fn command_basename(command: &str) -> &str {
@@ -1076,85 +747,6 @@ fn short_flag_contains_option(flag: &str, option: char) -> bool {
     flag.starts_with('-') && flag.chars().skip(1).any(|ch| ch == option)
 }
 
-fn is_safe_read_command(tokens: &[&str]) -> bool {
-    let head = tokens[0];
-
-    if matches!(
-        head,
-        "ls" | "pwd"
-            | "cd"
-            | "pushd"
-            | "popd"
-            | "dirs"
-            | "cat"
-            | "head"
-            | "tail"
-            | "wc"
-            | "rg"
-            | "grep"
-            | "which"
-            | "whereis"
-            | "basename"
-            | "dirname"
-            | "realpath"
-            | "readlink"
-            | "stat"
-            | "file"
-            | "du"
-            | "df"
-            | "cut"
-            | "tr"
-            | "sort"
-            | "uniq"
-            | "nl"
-            | "tree"
-            | "find"
-            | "echo"
-            | "printf"
-            | "env"
-            | "printenv"
-            | "id"
-            | "whoami"
-            | "uname"
-            | "date"
-            | "ps"
-            | "uptime"
-            | "history"
-            | "true"
-            | "false"
-            | "test"
-            | "["
-    ) {
-        return true;
-    }
-
-    if head == "sed" {
-        return is_sed_safe_read_command(tokens);
-    }
-
-    if is_python_query_command(tokens) {
-        return true;
-    }
-
-    if matches!(head, "tox" | "nox") && is_tool_query_command(tokens) {
-        return true;
-    }
-
-    if head == "command" {
-        return is_command_query(tokens);
-    }
-
-    if head == "builtin" {
-        return is_builtin_query(tokens);
-    }
-
-    if head == "git" {
-        return is_git_read_command(tokens);
-    }
-
-    false
-}
-
 fn is_sed_safe_read_command(tokens: &[&str]) -> bool {
     let mut saw_script = false;
     let mut index = 1;
@@ -1294,262 +886,6 @@ fn is_builtin_query(tokens: &[&str]) -> bool {
     tokens.get(1).copied().is_some_and(builtin_query_flag)
 }
 
-fn git_subcommand<'a>(tokens: &'a [&'a str]) -> Option<(usize, &'a str)> {
-    // Match on the basename so path-qualified git (`/usr/bin/git -C repo push`)
-    // is classified like bare `git`; otherwise a push/fetch misses network
-    // classification and runs with the sandbox network deny in force.
-    match tokens.first() {
-        Some(first) if command_basename(first) == "git" => {}
-        _ => return None,
-    }
-
-    let mut idx = 1;
-    while idx < tokens.len() {
-        let token = tokens[idx];
-        if token == "--" {
-            return tokens
-                .get(idx + 1)
-                .copied()
-                .map(|subcommand| (idx + 1, subcommand));
-        }
-        if !token.starts_with('-') {
-            return Some((idx, token));
-        }
-
-        let takes_value = matches!(
-            token,
-            "-c" | "-C"
-                | "--exec-path"
-                | "--git-dir"
-                | "--work-tree"
-                | "--namespace"
-                | "--super-prefix"
-                | "--config-env"
-        );
-        idx += 1;
-        if takes_value && !token.contains('=') && idx < tokens.len() {
-            idx += 1;
-        }
-    }
-
-    None
-}
-
-fn is_git_network_command(tokens: &[&str]) -> bool {
-    let Some((_, subcommand)) = git_subcommand(tokens) else {
-        return false;
-    };
-
-    matches!(
-        subcommand,
-        "clone" | "fetch" | "pull" | "push" | "ls-remote"
-    ) || is_git_remote_network(tokens)
-        || is_git_submodule_network(tokens)
-}
-
-fn is_git_read_command(tokens: &[&str]) -> bool {
-    let Some((_, subcommand)) = git_subcommand(tokens) else {
-        return true;
-    };
-
-    if subcommand == "submodule" {
-        return is_git_submodule_read(tokens);
-    }
-
-    match subcommand {
-        "status" | "diff" | "log" | "show" | "rev-parse" | "ls-files" | "ls-tree" | "blame"
-        | "grep" | "shortlog" | "describe" => true,
-        "branch" => is_git_branch_read(tokens),
-        "remote" => is_git_remote_read(tokens),
-        "tag" => is_git_tag_read(tokens),
-        _ => false,
-    }
-}
-
-fn is_git_branch_read(tokens: &[&str]) -> bool {
-    let Some((branch_idx, subcommand)) = git_subcommand(tokens) else {
-        return false;
-    };
-    if subcommand != "branch" {
-        return false;
-    }
-
-    const WRITE_FLAGS: &[&str] = &[
-        "-c",
-        "-C",
-        "-d",
-        "-D",
-        "-f",
-        "-m",
-        "-M",
-        "--copy",
-        "--delete",
-        "--move",
-        "--edit-description",
-        "--set-upstream-to",
-        "--track",
-        "--unset-upstream",
-    ];
-    if tokens
-        .iter()
-        .skip(branch_idx + 1)
-        .any(|token| WRITE_FLAGS.contains(token) || token.starts_with("--set-upstream-to="))
-    {
-        return false;
-    }
-
-    let list_mode = tokens
-        .iter()
-        .skip(branch_idx + 1)
-        .any(|token| matches!(*token, "-l" | "--list"));
-    let has_positional = tokens
-        .iter()
-        .skip(branch_idx + 1)
-        .any(|token| !token.starts_with('-'));
-
-    !has_positional || list_mode
-}
-
-fn git_remote_subcommand<'a>(tokens: &'a [&'a str]) -> Option<&'a str> {
-    let (remote_idx, subcommand) = git_subcommand(tokens)?;
-    if subcommand != "remote" {
-        return None;
-    }
-
-    tokens
-        .iter()
-        .skip(remote_idx + 1)
-        .find_map(|token| (!token.starts_with('-')).then_some(*token))
-}
-
-fn is_git_remote_network(tokens: &[&str]) -> bool {
-    let Some((remote_idx, subcommand)) = git_subcommand(tokens) else {
-        return false;
-    };
-    if subcommand != "remote" {
-        return false;
-    }
-
-    matches!(git_remote_subcommand(tokens), Some("show" | "update")) && !tokens.contains(&"-n")
-        || (matches!(git_remote_subcommand(tokens), Some("add"))
-            && tokens
-                .iter()
-                .skip(remote_idx + 1)
-                .any(|token| matches!(*token, "-f" | "--fetch")))
-}
-
-fn is_git_remote_read(tokens: &[&str]) -> bool {
-    let Some((_, subcommand)) = git_subcommand(tokens) else {
-        return false;
-    };
-    if subcommand != "remote" {
-        return false;
-    }
-
-    match git_remote_subcommand(tokens) {
-        None => true,
-        Some("get-url") => true,
-        Some("show") => tokens.contains(&"-n"),
-        _ => false,
-    }
-}
-
-fn is_git_tag_read(tokens: &[&str]) -> bool {
-    let Some((tag_idx, subcommand)) = git_subcommand(tokens) else {
-        return false;
-    };
-    if subcommand != "tag" {
-        return false;
-    }
-
-    const WRITE_FLAGS: &[&str] = &[
-        "-a",
-        "-d",
-        "-f",
-        "-m",
-        "-s",
-        "-u",
-        "--annotate",
-        "--delete",
-        "--force",
-        "--local-user",
-        "--message",
-        "--sign",
-    ];
-    if tokens
-        .iter()
-        .skip(tag_idx + 1)
-        .any(|token| WRITE_FLAGS.contains(token) || token.starts_with("--message="))
-    {
-        return false;
-    }
-
-    let read_flag = tokens.iter().skip(tag_idx + 1).any(|token| {
-        matches!(
-            *token,
-            "-l" | "-n"
-                | "-v"
-                | "--list"
-                | "--contains"
-                | "--merged"
-                | "--no-merged"
-                | "--points-at"
-                | "--sort"
-                | "--column"
-                | "--color"
-                | "--verify"
-        )
-    });
-    let has_positional = tokens
-        .iter()
-        .skip(tag_idx + 1)
-        .any(|token| !token.starts_with('-'));
-
-    !has_positional || read_flag
-}
-
-fn git_submodule_subcommand<'a>(tokens: &'a [&'a str]) -> Option<(usize, &'a str)> {
-    let (submodule_idx, subcommand) = git_subcommand(tokens)?;
-    if subcommand != "submodule" {
-        return None;
-    }
-
-    tokens
-        .iter()
-        .enumerate()
-        .skip(submodule_idx + 1)
-        .find_map(|(idx, token)| (!token.starts_with('-')).then_some((idx, *token)))
-}
-
-fn is_git_submodule_network(tokens: &[&str]) -> bool {
-    let Some((subcommand_idx, subcommand)) = git_submodule_subcommand(tokens) else {
-        return false;
-    };
-
-    match subcommand {
-        "update" => !tokens.contains(&"--no-fetch"),
-        "add" => tokens
-            .iter()
-            .skip(subcommand_idx + 1)
-            .any(|token| token.contains("://") || token.starts_with("git@")),
-        _ => false,
-    }
-}
-
-fn is_git_submodule_read(tokens: &[&str]) -> bool {
-    if tokens
-        .iter()
-        .any(|token| matches!(*token, "-h" | "--help" | "help"))
-    {
-        return true;
-    }
-
-    let Some((_, subcommand)) = git_submodule_subcommand(tokens) else {
-        return true;
-    };
-
-    matches!(subcommand, "status" | "summary")
-}
 impl Tool for BashTool {
     fn name(&self) -> &str {
         "bash"
